Do not use atomic blocks.
To cause a problem, the interrupt has to hit between the two STS
instructions that would write the two bytes to `gRfData`; the latch ISR
copies that to `data` to mutate.  This is exceedingly unlikely, but the
atomic block causes enough of a delay to be a real problem, more often.